Continued B/C refactoring
Another snapshot, since it seems functional now.
Improved grayscale toggling, as requested at https://forum.image.sc/t/feature-request-apply-classifiers-to-only-some-selected-objects/86383/8
