+++ qupath-gui-fx/src/main/java/qupath/lib/display/ImageDisplay.java
@@ -30,12 +30,16 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.stream.IntStream;
 
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -59,8 +63,6 @@ import qupath.lib.color.ColorTransformer;
 import qupath.lib.color.ColorTransformer.ColorTransformMethod;
 import qupath.lib.common.ColorTools;
 import qupath.lib.display.ChannelDisplayInfo.ModifiableChannelDisplayInfo;
-import qupath.lib.display.settings.ChannelSettings;
-import qupath.lib.display.settings.ImageDisplaySettings;
 import qupath.lib.gui.images.stores.AbstractImageRenderer;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.images.ImageData;
@@ -105,11 +107,22 @@ public class ImageDisplay extends AbstractImageRenderer {
 	
 	private ObjectBinding<ChannelDisplayMode> displayMode = Bindings.createObjectBinding(() -> calculateDisplayMode(),
 			useGrayscaleLutProperty(), useInvertedBackgroundProperty());
-	
-	private static transient Map<String, HistogramManager> cachedHistograms = Collections.synchronizedMap(new HashMap<>());
+
+	private static Map<String, HistogramManager> cachedHistograms = Collections.synchronizedMap(new HashMap<>());
 	private HistogramManager histogramManager = null;
 	
 	private static BooleanProperty showAllRGBTransforms = PathPrefs.createPersistentPreference("showAllRGBTransforms", true);
+
+	// Used to store the channel colors before switching to grayscale, so this can be restored later
+	// We use the names rather than the actual channels so that the colors are preserved if the image is changed,
+	// and we still do our best to preserve the matching channels
+	private transient Set<String> beforeGrayscaleChannels = new HashSet<>();
+
+	// Used to store the channel that should be selected when switching to grayscale (optional).
+	// This is useful to develop more intuitive interfaces & prevent surprises when switching to grayscale mode.
+	// For example, this could be in the selected item in a list or table (independent of whether the channel is
+	// shown or hidden).
+	private transient ObjectProperty<ChannelDisplayInfo> switchToGrayscaleChannel = new SimpleObjectProperty<>();
 
 	/**
 	 * Constructor.
@@ -119,6 +132,11 @@ public class ImageDisplay extends AbstractImageRenderer {
 		setImageData(imageData, false);
 		useGrayscaleLuts.addListener((v, o, n) -> {
 			if (n) {
+				// Snapshot the names of channels active before switching to grayscale
+				selectedChannels.stream().map(ChannelDisplayInfo::getName).forEach(beforeGrayscaleChannels::add);
+				var swithToGrayscale = switchToGrayscaleChannel.get();
+				if (swithToGrayscale != null)
+					setChannelSelected(swithToGrayscale, true);
 				if (lastSelectedChannel != null)
 					setChannelSelected(lastSelectedChannel, true);
 				else if (!selectedChannels.isEmpty())
@@ -126,6 +144,15 @@ public class ImageDisplay extends AbstractImageRenderer {
 				else if (availableChannels.isEmpty()) {
 					setChannelSelected(availableChannels.get(0), true);
 				}
+			} else {
+				// Restore the channels that were active before switching to grayscale, if possible
+				if (!beforeGrayscaleChannels.isEmpty()) {
+					var channelsToSelect = availableChannels().stream().filter(c -> beforeGrayscaleChannels.contains(c.getName())).toList();
+					if (!channelsToSelect.isEmpty()) {
+						selectedChannels.setAll(channelsToSelect);
+					}
+				}
+				beforeGrayscaleChannels.clear();
 			}
 			saveChannelColorProperties();
 		});
@@ -336,8 +363,39 @@ public class ImageDisplay extends AbstractImageRenderer {
 	 */
 	public void refreshChannelOptions() {
 		updateChannelOptions(false);
-		
 	}
+
+	/**
+	 * Property indicating which channel should be used if {@link #useGrayscaleLutProperty()} is turned on.
+	 * This is useful to develop more intuitive interfaces & prevent surprises when switching to grayscale mode.
+	 * <p>
+	 * Settings this value does not have any immediate effect on whether channels are selected or not, but rather it
+	 * is only used when switching to grayscale mode.
+	 * @return
+	 * @since v0.5.0
+	 */
+	public ObjectProperty<ChannelDisplayInfo> switchToGrayscaleChannelProperty() {
+		return switchToGrayscaleChannel;
+	}
+
+	/**
+	 * Set the value of {@link #switchToGrayscaleChannelProperty()}.
+	 * @param channel
+	 * @since v0.5.0
+	 */
+	public void setSwitchToGrayscaleChannel(ChannelDisplayInfo channel) {
+		switchToGrayscaleChannel.set(channel);
+	}
+
+	/**
+	 * Get the value of {@link #switchToGrayscaleChannelProperty()}.
+	 * @return
+	 * @since v0.5.0
+	 */
+	public ChannelDisplayInfo getSwitchToGrayscaleChannel() {
+		return switchToGrayscaleChannel.get();
+	}
+
 
 	private void updateChannelOptions(boolean serverChanged) {
 		
@@ -352,6 +410,8 @@ public class ImageDisplay extends AbstractImageRenderer {
 		if (server == null) {
 			selectedChannels.clear();
 			channelOptions.clear();
+			if (!switchToGrayscaleChannel.isBound())
+				switchToGrayscaleChannel.set(null);
 			return;
 		}
 		
@@ -412,6 +472,17 @@ public class ImageDisplay extends AbstractImageRenderer {
 		}
 		channelOptions.setAll(tempChannelOptions);
 		selectedChannels.setAll(tempSelectedChannels);
+
+		// Attempt to preserve the switchToGrayscaleChannel, if possible
+		if (!switchToGrayscaleChannel.isBound()) {
+			var switchToGrayscale = switchToGrayscaleChannel.get();
+			if (switchToGrayscale != null) {
+				if (!channelOptions.contains(switchToGrayscale))
+					switchToGrayscaleChannel.set(
+							channelOptions.stream().filter(c -> Objects.equals(c.getName(), switchToGrayscale.getName())).findFirst().orElse(null)
+					);
+			}
+		}
 	}
